Unit test cleanup
