+++ test/GiGraph.Dot.Output.Tests/DotGraphSaveToTextWriterTest.cs
@@ -1,9 +1,10 @@
+using System.Text;
 using GiGraph.Dot.Extensions;
 using Xunit;
 
 namespace GiGraph.Dot.Output.Tests;
 
-public partial class DotGraphSaveTest
+public class DotGraphSaveToTextWriterTest
 {
     [Fact]
     public void graph_is_saved_to_text_writer_without_flush()
@@ -22,5 +23,12 @@ public partial class DotGraphSaveTest
 
         writer.Flush();
         Assert.Equal(dotString.Length, stream.Length);
+    }
+
+    private static StreamWriter CreateStreamWriter(Stream stream, Encoding? encoding, int? bufferSize = null)
+    {
+        // this is for compatibility with .NET Standard where encoding has to be specified if you want to set leaveOpen
+        encoding ??= new UTF8Encoding(encoderShouldEmitUTF8Identifier: false, throwOnInvalidBytes: true);
+        return new StreamWriter(stream, encoding, bufferSize: bufferSize ?? -1, leaveOpen: true);
     }
 }
