+++ test/GiGraph.Dot.Output.Tests/EncodingHelpers/EncodingHelper.cs
@@ -1,0 +1,18 @@
+using System.Text;
+using Xunit;
+
+namespace GiGraph.Dot.Output.Tests.EncodingHelpers;
+
+internal static class EncodingHelper
+{
+    public static bool HasBom(Stream stream)
+    {
+        var bom = Encoding.UTF8.GetPreamble();
+        Assert.NotEmpty(bom);
+
+        var buffer = new byte[bom.Length];
+        stream.ReadExactly(buffer, 0, buffer.Length);
+
+        return buffer.SequenceEqual(bom);
+    }
+}
